+++ gr-blocks/include/gnuradio/blocks/peak_detector2_fb.h
@@ -1,6 +1,6 @@
 /* -*- c++ -*- */
 /*
- * Copyright 2007,2013 Free Software Foundation, Inc.
+ * Copyright 2007,2013,2015 Free Software Foundation, Inc.
  *
  * This file is part of GNU Radio
  *
@@ -36,7 +36,7 @@ namespace gr {
      * \details
      * If a peak is detected, this block outputs a 1, or it outputs
      * 0's. A separate debug output may be connected, to view the
-     * internal EWMA described below.
+     * internal estimated mean described below.
      */
     class BLOCKS_API peak_detector2_fb : virtual public sync_block
     {
@@ -47,13 +47,21 @@ namespace gr {
       /*!
        * Build a peak detector block with float in, byte out.
        *
-       * \param threshold_factor_rise The threshold factor determins
-       *        when a peak is present. An EWMA average of the signal is
-       *        calculated and when the value of the signal goes over
-       *        threshold_factor_rise*average, we call the peak.
+       * \param threshold_factor_rise The threshold factor determines
+       *        when a peak is present. An average of the input signal
+       *        is calculated (through a single-pole autoregressive
+       *        filter) and when the value of the input signal goes
+       *        over threshold_factor_rise*average, we assume we are
+       *        in the neighborhood of a peak. The block will then
+       *        find the position of the maximum within a window of
+       *        look_ahead samples starting at the point where the
+       *        threshold was crossed upwards.
        * \param look_ahead The look-ahead value is used when the
-       *        threshold is found to locate the peak within this range.
-       * \param alpha The gain value of a single-pole moving average filter.
+       *        threshold is crossed upwards to locate the peak within
+       *        this range.
+       * \param alpha One minus the pole of a single-pole
+       *        autoregressive filter that evaluates the average of
+       *        the input signal.
        */
       static sptr make(float threshold_factor_rise=7,
                        int look_ahead=1000, float alpha=0.001);
@@ -71,7 +79,7 @@ namespace gr {
       /*! \brief Set the running average alpha
        *  \param alpha new alpha for running average
        */
-      virtual void set_alpha(int alpha) = 0;
+      virtual void set_alpha(float alpha) = 0;
 
       /*! \brief Get the threshold factor value for the rise time
        *  \return threshold factor
